Introduce game process folder
