+++ Assets/Scripts/BorderLogic.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e355709a8279e4c45be21195c1dfe123
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
