+++ Assets/Scripts/GameProcess/HandDrower.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5ff491de53bfc407faf4763d113b241d
+guid: 84034b9db1f654c149b438c775bf7560
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
