+++ Assets/Scripts/GameProcess/SimpleCard.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f7c780b52ed5d43e180f00488830fb3c
+guid: 1a65ea3c2b33e4e688c74964b1b533a3
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
