+++ Assets/Scripts/Interactions.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aab6e9eb031894cd5abd6876d36325f8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
